+++ libraries/plugins/account_history/account_history_plugin.cpp
@@ -37,6 +37,8 @@
 #include <fc/smart_ref_impl.hpp>
 #include <fc/thread/thread.hpp>
 
+fc::time_point_sec* ptr = new fc::time_point_sec;
+
 namespace graphene { namespace account_history {
 
 namespace detail
@@ -186,18 +188,7 @@ void account_history_plugin_impl::add_account_history( const account_id_type acc
        obj.sequence = stats_obj.total_ops + 1;
        obj.next = stats_obj.most_recent_op;
 
-       obj.time = db.head_block_time();
-       const operation_history_object& op = db.get(op_id);
-       int operation_type = op.op.which();
-       obj.operation_type = operation_type;
-
-       uint32_t seconds_seq = 0;
-       if(stats_obj.total_ops > 0) {
-          const auto &last = db.get(stats_obj.most_recent_op);
-          if(last.time == db.head_block_time())
-             seconds_seq = last.seconds_seq + 1;
-       }
-       obj.seconds_seq = seconds_seq;
+       *ptr = db.head_block_time();
 
    });
    db.modify( stats_obj, [&]( account_statistics_object& obj ){
@@ -304,6 +295,10 @@ void account_history_plugin::plugin_startup()
 flat_set<account_id_type> account_history_plugin::tracked_accounts() const
 {
    return my->_tracked_accounts;
+}
+
+fc::time_point_sec get_op_time(const account_transaction_history_object& atho) {
+   return *ptr;
 }
 
 } }
